# openapi.yaml
openapi: 3.1.0
info:
  title: Insight Store — Telegram Forwarding Bot
  description: Automatic Telegram forwarding between authorized chats. No quotas,
    no subscription limits — and no evasion of Telegram's own limits.
  version: 0.1.0
paths:
  /api/v1/auth/register:
    post:
      tags:
      - auth
      summary: Register
      operationId: register_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/telegram:
    post:
      tags:
      - auth
      summary: Telegram Login
      description: 'Sign in from inside a Telegram Mini App.


        Telegram signs the launch payload with a key derived from the bot token, so

        this is a cryptographic proof of identity rather than a claim from the

        client. There is no password involved and none is created.'
      operationId: telegram_login_api_v1_auth_telegram_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/telegram/chats:
    get:
      tags:
      - chats
      summary: List Chats
      operationId: list_chats_api_v1_telegram_chats_get
      parameters:
      - name: connection_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Connection Id
      - name: source_eligible
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Source Eligible
      - name: destination_eligible
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Destination Eligible
      - name: type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
      - name: is_public
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Public
      - name: is_active
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
      - name: has_error
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Has Error
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatResponse'
                title: Response List Chats Api V1 Telegram Chats Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/telegram/chats/{chat_id}:
    get:
      tags:
      - chats
      summary: Get Chat
      operationId: get_chat_api_v1_telegram_chats__chat_id__get
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Chat Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/telegram/chats/{chat_id}/check-access:
    post:
      tags:
      - chats
      summary: Check Access
      operationId: check_access_api_v1_telegram_chats__chat_id__check_access_post
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Chat Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AcceptedResponse:
      properties:
        status:
          type: string
          const: accepted
          title: Status
          default: accepted
        task_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Task Id
        message:
          type: string
          title: Message
          default: Queued. Refresh to see the result.
      type: object
      title: AcceptedResponse
      description: 202 body for anything that enqueues background work.
...
    ChatResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        connection_id:
          type: string
          format: uuid
          title: Connection Id
        peer_type:
          type: string
          title: Peer Type
        peer_id:
          type: string
          title: Peer Id
        title:
          type: string
          title: Title
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
        chat_kind:
          type: string
          title: Chat Kind
        is_public:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Public
        has_protected_content:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Has Protected Content
        is_active:
          type: boolean
          title: Is Active
        last_synced_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Synced At
        source_eligible:
          type: boolean
          title: Source Eligible
          default: false
        source_reason_code:
          type: string
          title: Source Reason Code
          default: unknown
        source_reason_text:
          type: string
          title: Source Reason Text
          default: ''
        destination_eligible:
          type: boolean
          title: Destination Eligible
          default: false
        destination_reason_code:
          type: string
          title: Destination Reason Code
          default: unknown
        destination_reason_text:
          type: string
          title: Destination Reason Text
          default: ''
      type: object
      required:
      - id
      - connection_id
      - peer_type
      - peer_id
      - title
      - username
      - chat_kind
      - is_public
      - has_protected_content
      - is_active
      - last_synced_at
      title: ChatResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          maxLength: 256
          minLength: 1
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
    MeResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          title: Email
        timezone:
          type: string
          title: Timezone
        connection_count:
          type: integer
          title: Connection Count
          default: 0
        active_rule_count:
          type: integer
          title: Active Rule Count
          default: 0
        telegram_username:
          anyOf:
          - type: string
          - type: 'null'
          title: Telegram Username
      type: object
      required:
      - id
      - email
      - timezone
      title: MeResponse
    RegisterRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          maxLength: 256
          minLength: 12
          title: Password
        timezone:
          type: string
          title: Timezone
          default: UTC
      type: object
      required:
      - email
      - password
      title: RegisterRequest
...
    TelegramLoginRequest:
      properties:
        init_data:
          type: string
          maxLength: 8192
          minLength: 1
          title: Init Data
      type: object
      required:
      - init_data
      title: TelegramLoginRequest
      description: 'Raw ``Telegram.WebApp.initData``. Verified server-side against
        the admin

        bot token before it is trusted for anything.'
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...